refactor: move lazyAssert function to utils/assert.ts
